+++ src/netprotocols/layer2/__init__.py
@@ -1,4 +1,5 @@
 from netprotocols.layer2.arp import ARP
 from netprotocols.layer2.ethernet import Ethernet
+from netprotocols.layer2.vlan import VLAN
 
-__all__ = ["ARP", "Ethernet"]
+__all__ = ["ARP", "VLAN", "Ethernet"]